Reverted change of cloudFromStereoImages() decimation, should be done after StereoBM to avoid backward compatibility problems

+++ corelib/include/rtabmap/core/util3d.h
@@ -129,7 +129,7 @@ pcl::PointCloud<pcl::PointXYZRGB>::Ptr RTABMAP_EXP cloudFromStereoImages(
 		const cv::Mat & imageLeft,
 		const cv::Mat & imageRight,
 		const StereoCameraModel & model,
-		float decimation = 1.0f,
+		int decimation = 1,
 		float maxDepth = 0.0f,
 		float minDepth = 0.0f,
 		std::vector<int> * validIndices = 0,
